# packages/api-contract/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/spec-details": {
      "get": {
        "operationId": "SpecDetailsController_findAll",
        "parameters": [
          {
            "name": "filters",
            "required": false,
            "in": "query",
            "description": "Refine filters JSON",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "pagination",
            "required": false,
            "in": "query",
            "description": "Refine pagination JSON",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "返回规格明细列表和总数",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpecDetailListResponseDto"
                }
              }
            }
          }
        },
        "summary": "获取规格明细列表",
        "tags": [
          "Spec Details"
        ]
      },
      "post": {
        "operationId": "SpecDetailsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSpecDetailDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpecDetailResponseDto"
                }
              }
            }
          }
        },
        "summary": "创建规格明细",
        "tags": [
          "Spec Details"
        ]
      }
    },
    "/api/spec-details/{id}": {
      "get": {
        "operationId": "SpecDetailsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "返回规格明细详情",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpecDetailResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "规格明细不存在"
          }
        },
        "summary": "获取规格明细详情",
        "tags": [
          "Spec Details"
        ]
      },
      "put": {
        "operationId": "SpecDetailsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSpecDetailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpecDetailResponseDto"
                }
              }
            }
          }
        },
        "summary": "更新规格明细",
        "tags": [
          "Spec Details"
        ]
      },
      "delete": {
        "operationId": "SpecDetailsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功"
          }
        },
        "summary": "删除规格明细",
        "tags": [
          "Spec Details"
        ]
      }
    },
    "/api/sizes": {
      "get": {
        "operationId": "SizesController_findAll",
        "parameters": [
          {
            "name": "filters",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "pagination",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "返回尺码列表和总数",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SizeListResponseDto"
                }
              }
            }
          }
        },
        "summary": "获取尺码列表",
        "tags": [
          "Sizes"
        ]
      },
      "post": {
        "operationId": "SizesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSizeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SizeResponseDto"
                }
              }
            }
          }
        },
        "summary": "创建尺码",
        "tags": [
          "Sizes"
        ]
      }
    },
    "/api/sizes/{id}": {
      "get": {
        "operationId": "SizesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "返回尺码详情",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SizeResponseDto"
                }
              }
            }
          }
        },
        "summary": "获取尺码详情",
        "tags": [
          "Sizes"
        ]
      },
      "put": {
        "operationId": "SizesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSizeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SizeResponseDto"
                }
              }
            }
          }
        },
        "summary": "更新尺码",
        "tags": [
          "Sizes"
        ]
      },
      "delete": {
        "operationId": "SizesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功"
          }
        },
        "summary": "删除尺码",
        "tags": [
          "Sizes"
        ]
      }
    },
...
  },
  "info": {
    "title": "SpecMaster API",
    "description": "服装配方管理系统后端 API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "SpecDetailResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "规格明细ID",
            "example": 1
          },
          "bomItemId": {
            "type": "number",
            "description": "BOM配料ID",
            "example": 1
          },
          "size": {
            "type": "string",
            "description": "尺码",
            "example": "M"
          },
          "specValue": {
            "type": "string",
            "description": "规格值",
            "example": "5.5"
          },
          "specUnit": {
            "type": "string",
            "description": "规格单位",
            "example": "cm"
          },
          "sortOrder": {
            "type": "number",
            "description": "排序顺序",
            "example": 0
          },
          "createdAt": {
            "type": "string",
            "description": "创建时间"
          },
          "updatedAt": {
            "type": "string",
            "description": "更新时间"
          }
        },
        "required": [
          "id",
          "bomItemId",
          "specValue",
          "specUnit",
          "sortOrder",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CreateSpecDetailDto": {
        "type": "object",
        "properties": {
          "bomItemId": {
            "type": "number",
            "description": "所属配料明细ID"
          },
          "size": {
            "type": "string",
            "description": "尺码",
            "example": "M"
          },
          "specValue": {
            "type": "string",
            "description": "规格值",
            "example": "58.5"
          },
          "specUnit": {
            "type": "string",
            "description": "规格单位",
            "example": "cm"
          },
          "sortOrder": {
            "type": "number",
            "description": "排序顺序"
          }
        },
        "required": [
          "bomItemId",
          "specValue",
          "specUnit"
        ]
      },
...
      "SpecDetailListResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpecDetailResponseDto"
            }
          },
          "total": {
            "type": "number",
            "description": "总记录数"
          }
        },
        "required": [
          "data",
          "total"
        ]
      },
      "UpdateSpecDetailDto": {
        "type": "object",
        "properties": {
          "bomItemId": {
            "type": "number",
            "description": "所属配料明细ID"
          },
          "size": {
            "type": "string",
            "description": "尺码",
            "example": "M"
          },
          "specValue": {
            "type": "string",
            "description": "规格值",
            "example": "58.5"
          },
          "specUnit": {
            "type": "string",
            "description": "规格单位",
            "example": "cm"
          },
          "sortOrder": {
            "type": "number",
            "description": "排序顺序"
          }
        }
      },
      "SizeResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "尺码ID",
            "example": 1
          },
          "sizeCode": {
            "type": "string",
            "description": "尺码代码",
            "example": "M"
          },
          "sizeName": {
            "type": "string",
            "description": "尺码名称",
            "example": "中号"
          },
          "sortOrder": {
            "type": "number",
            "description": "排序顺序",
            "example": 0
          },
          "note": {
            "type": "string",
            "description": "备注"
          },
          "isActive": {
            "type": "boolean",
            "description": "是否启用",
            "example": true
          },
          "createdAt": {
            "type": "string",
            "description": "创建时间"
          },
          "updatedAt": {
            "type": "string",
            "description": "更新时间"
          }
        },
        "required": [
          "id",
          "sizeCode",
          "sizeName",
          "sortOrder",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
      "SizeListResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SizeResponseDto"
            }
          },
          "total": {
            "type": "number",
            "description": "总记录数"
          }
        },
        "required": [
          "data",
          "total"
        ]
      },
      "CreateSizeDto": {
        "type": "object",
        "properties": {
          "sizeCode": {
            "type": "string",
            "description": "尺码代码",
            "example": "M"
          },
          "sizeName": {
            "type": "string",
            "description": "尺码名称",
            "example": "中号"
          },
          "sortOrder": {
            "type": "number",
            "description": "排序顺序"
          },
          "note": {
            "type": "string",
            "description": "备注"
          },
          "isActive": {
            "type": "boolean",
            "description": "是否启用",
            "default": true
          }
        },
        "required": [
          "sizeCode",
          "sizeName"
        ]
      },
      "UpdateSizeDto": {
        "type": "object",
        "properties": {
          "sizeCode": {
            "type": "string",
            "description": "尺码代码"
          },
          "sizeName": {
            "type": "string",
            "description": "尺码名称"
          },
          "sortOrder": {
            "type": "number",
            "description": "排序顺序"
          },
          "note": {
            "type": "string",
            "description": "备注"
          },
          "isActive": {
            "type": "boolean",
            "description": "是否启用"
          }
        }
      },
...
    }
  }
}